server/requirements.txt
autopep8==1.5.4
cachelib==0.1.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.1.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
Werkzeug==1.0.1
